Refactor efficientsam module
